Fix QS expansion animations
* Fix gap between tiles when expanding
* Fix second page collapsing animation: tiles now will shrink and at the
end be replaced by QQS tiles. We need to create the animators as we move
through the pages, as PagedTileLayout only has the current page and the
(up to) two adjacent ones. This means that if the page has not been
added, we don't have correct positions and measures.

See bug for video
Fixes: 207895040
Test: manual

Change-Id: Id2b14b11ace4199ed07f7675b6575cb4b71e93d5
Signed-off-by: strwbrry143 <alexfinhart@gmail.com>