Smooth out the alpha / colour animation for drag to split
Previously there were 2 animating surfaces, one for the highlight and
one for the splashscreen. When they animated together there was a bit
of undesirable crossfading.

This change makes it so that the colours and alpha animate together via
ArgbEvaluator and are only applied to 1 surface (the background color
drawable).

It also seems like you can continue to get drag events even after a
drop, so if a drop has happened skip animating. When a drag is
initiated again then mHasDropped is reset again.

Test: manual - drag an app into split and move it between the two split
               dropzones, note that the colour animation is smooth.
             - wildly drag between the two dropzones & drop, check that
               on the next drag that animations animate from alpha 0 rather
               than jump cutting to an alpha > 0.
Bug: 207011867
Change-Id: I8a5a681530d889c5c6ef4ad3a0b31af17998af73
Signed-off-by: strwbrry143 <alexfinhart@gmail.com>