Improve compaction to abort when system changes to awake state
While the screen is off we sometimes issue compactions as they are not
disruptive to the user. However, it is possible to wake up and have
compactions running or scheduled so this may lead to jank. This patch
cancels any pending or running compactions that were scheduled to happen
during this non interactive type when the system becomes awake to avoid
jank.

Test: Manual
Bug: 214654755
Change-Id: Ie9b2cbaa7093d63e77f8666d1e6f050ab5610a1d
Signed-off-by: strwbrry143 <alexfinhart@gmail.com>