Fixing expansion state in NotificationPanelViewController
This also reverts ag/16835086, which was just a quick hack while this is a proper long-term solution.
The main change here is making sure mQsExpanded flag accurately represents the state in split shade. To do that, mQsExpandImmediately is reused and it makes sure that QS ends up in proper state once shade is expanded.
That cascaded to other changes/fixes, one of the major one being changing mQsExpanded in NotificationStackScrollLayout to mQsFullScreen - otherwise every use of mQsExpanded would have to be updated with checking if we’re in split shade mode. mQsFullScreen better represents change in notifications behaviour depending on the state.

Fixes: 218268829
Test: Expand split shade and check all tiles - especially on the second page - are working correctly
Change-Id: Id65c73c034fb8705eed8274416f125e4d861ac30
str: adapted to 12L - central surfaces - statusbar
Signed-off-by: str143 <alexfinhart@gmail.com>