Fix drag to split in RTL
The code assumes that the views are in LTR order, so when the device
is in RTL, the wrong view gets highlighted. The code thinks dropzone1
is on the left side, but in RTL it'd be on the right.

Fix for this is to just always layout the views in LTR.

Bug: 228530510
Test: manual - put the device in RTL and drag an app into split
             => verify that the highlight is on the correct side
Change-Id: I05e548143ed7042c4151252b46fa1b96d8ac7409
Signed-off-by: strwbrry143 <alexfinhart@gmail.com>